+++ db/tasks.php
@@ -15,7 +15,7 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Plugin version and other metadata.
+ * Scheduled task registrations.
  *
  * @package    local_groupdist
  * @copyright  2026 Anderson Blaine
@@ -24,9 +24,14 @@
 
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->component = 'local_groupdist';
-$plugin->version = 2026081305;
-$plugin->requires = 2025100600;
-$plugin->maturity = MATURITY_ALPHA;
-$plugin->release = 'v0.1.0';
-$plugin->supported = [501, 502];
+$tasks = [
+    [
+        'classname' => 'local_groupdist\task\cleanup_audit',
+        'blocking' => 0,
+        'minute' => 'R',
+        'hour' => '3',
+        'day' => '*',
+        'month' => '*',
+        'dayofweek' => '*',
+    ],
+];